+++ Backend/DataAccessLayer/DTOs/DTO.cs
@@ -10,7 +10,7 @@ namespace IntroSE.Kanban.Backend.DataAccessLayer
     {
         public const string IDColumnName = "id";
         protected DalController _controller;
-        public long id { get; set; } = -1;
+        public int id { get; set; } = -1;
         protected DTO(DalController controller)
         {
             _controller = controller;
